Abins: unify 1D and 2D powder intensity evaluation
Implement concept of 'angles' consistently over direct and
indirect-geometry instruments, and we can remove the block that treats
them differently.

For now this is more 'stateful' than necessary, but demonstrates that
tests pass without API changes.
