+++ scripts/abins/instruments/toscainstrument.py
@@ -6,13 +6,14 @@
 # SPDX - License - Identifier: GPL - 3.0 +
 import abins
 import abins.parameters
-from abins.constants import WAVENUMBER_TO_INVERSE_A
 from .indirectinstrument import IndirectInstrument
 
 
 class ToscaInstrument(IndirectInstrument):
     """
-    Class for TOSCA and TOSCA-like instruments.
+    ISIS TOSCA instrument
+
+    A quadratic model is used for energy resolution
     """
     parameters = abins.parameters.instruments['TOSCA']
 
@@ -21,30 +22,6 @@ class ToscaInstrument(IndirectInstrument):
 
         if self._setting != '':
             print('WIP: TOSCA Instrument settings not yet implemented')
-
-    @classmethod
-    def calculate_q_powder(cls, input_data=None):
-        """Calculates squared Q vectors for TOSCA and TOSCA-like instruments.
-
-        By the cosine law Q^2 = k_f^2 + k_i^2 - 2 k_f k_i cos(theta)
-
-        where k are determined from
-        abins.parameters.instruments['TOSCA']['final_neutron_energy']
-        and the input series of vibrational frequencies and cos(theta) is
-        precomputed as abins.parameters.instruments['TOSCA']['cos_scattering_angle']
-
-        :param input_data:
-            frequencies (in cm-1) which should be used to construct Q2
-
-        :returns:
-            Q^2 array (in cm-1) corresponding to input frequencies,
-            constrained by conservation of mass/momentum and TOSCA geometry
-        """
-
-        k2_i = (input_data + cls.parameters['final_neutron_energy']) * WAVENUMBER_TO_INVERSE_A
-        k2_f = cls.parameters['final_neutron_energy'] * WAVENUMBER_TO_INVERSE_A
-        result = k2_i + k2_f - 2 * (k2_i * k2_f) ** 0.5 * cls.parameters['cos_scattering_angle']
-        return result
 
     @classmethod
     def get_sigma(cls, frequencies):
